adding alerts and dialogs after CRUD operations

+++ client/src/components/Employees.js
@@ -1,11 +1,13 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { EditIcon, DeleteIcon } from "@chakra-ui/icons";
-import Alert from "./Alert";
-import { Button, Input, Select } from "@chakra-ui/react";
+import AlertDialogExample from "./AlertDialogExample";
+import { Button, Input, Select, Alert, AlertIcon } from "@chakra-ui/react";
 const Employees = ({ departments }) => {
   const [employees, setEmployees] = useState([]);
   const [editRow, setEditRow] = useState(-1);
+  const [alertUpdate, setAlertUpdate] = useState(false);
+  const [alertDelete, setAlertDelete] = useState(false);
   const [employee, setEmployee] = useState({
     name: "",
     surname: "",
@@ -23,13 +25,58 @@ const Employees = ({ departments }) => {
   const deleteEmployee = (id) => {
     axios
       .delete(`http://localhost:3500/employee/${id}`)
-      .then((res) => alert(res.data.message))
+      .then((res) => {
+        setAlertDelete(true);
+        setTimeout(() => setAlertDelete(false), 2000);
+      })
       .catch((err) => alert(err.message));
   };
-  const updateEmployee = (id) => {};
+  const handleEdit = (e, i) => {
+    const id = departments.find((dep) => dep.name === e.departementName)._id;
+    setEmployee({
+      name: e.name,
+      surname: e.surname,
+      departementId: id,
+    });
+    setEditRow(i);
+  };
+  const handleDepartmentId = (e) => {
+    const id = departments.find((dep) => dep.name === e.target.value)._id;
+    setEmployee((prev) => ({
+      ...prev,
+      departementId: id,
+    }));
+  };
+  const updateEmployee = (id) => {
+    axios
+      .put(`http://localhost:3500/employee/${id}`, employee)
+      .then((res) => {
+        setAlertUpdate(true);
+        setTimeout(() => setAlertUpdate(false), 2000);
+      })
+      .catch((err) => alert(err.message));
+  };
 
   return (
     <div className="employees">
+      {alertDelete && (
+        <Alert
+          status="warning"
+          style={{ width: "80%", margin: "1% 5% 1% 5% " }}
+        >
+          <AlertIcon />
+          employee deleted successfully
+        </Alert>
+      )}
+      {alertUpdate && (
+        <Alert
+          status="success"
+          style={{ width: "80%", margin: "1% 5% 1% 5% " }}
+        >
+          <AlertIcon />
+          updated successfully
+        </Alert>
+      )}
       <h3>list of employees</h3>
       {!employees.length ? (
         <h2>loading ....</h2>
@@ -52,21 +99,40 @@ const Employees = ({ departments }) => {
               >
                 <td>
                   {editRow === i ? (
-                    <Input defaultValue={e.name} />
+                    <Input
+                      defaultValue={e.name}
+                      onChange={(event) =>
+                        setEmployee((prev) => ({
+                          ...prev,
+                          name: event.target.value,
+                        }))
+                      }
+                    />
                   ) : (
                     <span>{e.name}</span>
                   )}
                 </td>
                 <td>
                   {editRow === i ? (
-                    <Input defaultValue={e.surname} />
+                    <Input
+                      defaultValue={e.surname}
+                      onChange={(event) =>
+                        setEmployee((prev) => ({
+                          ...prev,
+                          surname: event.target.value,
+                        }))
+                      }
+                    />
                   ) : (
                     <span>{e.surname}</span>
                   )}
                 </td>
                 <td>
                   {editRow === i ? (
-                    <Select defaultValue={e.departementName}>
+                    <Select
+                      defaultValue={e.departementName}
+                      onChange={handleDepartmentId}
+                    >
                       {departments.map((el) => (
                         <option value={el.name} key={el._id}>
                           {el.name}
@@ -82,7 +148,7 @@ const Employees = ({ departments }) => {
                     width: "10%",
                   }}
                 >
-                  <Alert
+                  <AlertDialogExample
                     text={<DeleteIcon />}
                     action={() => deleteEmployee(e.id)}
                   />
@@ -94,10 +160,12 @@ const Employees = ({ departments }) => {
                     justifyContent: "space-evenly",
                   }}
                 >
-                  <button onClick={() => setEditRow(i)}>
+                  <button onClick={() => handleEdit(e, i)}>
                     <EditIcon />
                   </button>
-                  <Button>Save Update</Button>
+                  <Button onClick={() => updateEmployee(e.id)}>
+                    Save Update
+                  </Button>
                 </td>
               </tr>
             ))}
